Fix notifications trigger when nothing executed
Notifications ran even when no command was executed, e. g. when
cancelling current prompt with ^C or confirming the command (pressing
Enter) when there is an empty string so nothing gets executed.

Now we check for `$1` variable provided by `preexec` hook to check if
there actually was any command executed to avoid million of annoying popups.